+++ src/interpreter/eval.cpp
@@ -13,7 +13,6 @@
 #include "interpreter.hpp"
 #include "utils.hpp"
 #include "value.hpp"
-#include "value_as.hpp"
 
 namespace Interpreter {
